Fix all edit-clear-location-ltr icons
icons with -ltr suffix should have the icon going from left to right and
are used (counterintuitively) in situations where the text layout is
right to left, if the normal icon points rtl.

This patch changes the icons from being a symlink to the same icon to being a copy mirrored in inkscape.

BUG: 366519
REVIEW: 128694