CMakeLists.txt - Do not find bash on Windows
On Windows, even if we have bash (which is provided, for example,
by the GoW project) we don't have the tools required by validate-svg.sh
to actually run. We'd need to also ensure the unixy 'find' command
(not the Windows 'find' command) as well as xmllint.

For now let's not even bother on Windows.

A better solution is to write a new portable script in Python
and also you need to make sure xmllint is available.