change 32px icons for playlist shuffle and repeat
Summary:
the 32px icons for media-playlist-shuffle and media-playlist-repeat look totally different than the icons with sizes <= 24px.
Also, they do not fit with the current design with their sharp corners. Change to the design of the smaller icons. Icons were created using a scaled version of the 24px icons.

Reviewers: #breeze, #vdg, andreaska, andreask, ngraham

Reviewed By: #vdg, andreask, ngraham

Subscribers: bcooksley, andreask, ngraham, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10292