rename 64 px icons added for elisa
Summary:
see D10492. The new 64px icons with the same name cause issues because the icons now look totally different with different sizes.
rename the files so that explicit support must be added in Elisa. the view-media-playlist icon is actually the JuK icon, and should stay unique, so delete it again

Reviewers: #breeze, #vdg, andreask

Reviewed By: andreask

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D11461