ScalableTest, add "scalable" plasma-browser-integration
With the introduction of the plasma-browser-integration icon (Add
plasma-browser-integration system tray icon, 47b8e71a) the ScalableTest started
failing [1]. The test seems to expect a "larger" icon (greater than 48)
available to consider that there is a scalable icon. I'm not
sure if this test is still useful, as all the icons are now in svg format, and
while they might be designed for a low resolution, they are "scalable".

This patch simply adds symlinks to the 48 and 64 res, in order to make the
test pass.

[1]: https://build.kde.org/job/Frameworks%20breeze-icons%20kf5-qt5%20SUSEQt5.10/81/

BUG: 393999
FIXED-IN: 5.49