Fix contrast and other small visual issues of Kdots icon
Summary:
Fix contrast with Breeze light color scheme as well small issues with the lines connecting the dots in the icon.

Before:
{F6417532}
The blue line at the bottom is above the blue dot when it should be below. The line at the top does not reach far enough under the dot, a small gap is visible.

After:
{F6418103}
The contrast with a Breeze Light background color is much better. The blue lines are now all one path and are properly below the blue dots. They all reach under the blue dots. The right red dot has a shadow. All shadows are now above the connecting lines.

Reviewers: #vdg, ndavis, ngraham

Reviewed By: #vdg, ndavis, ngraham

Subscribers: ngraham, ndavis, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D16840