Add a proper python-bytecode icon, use consistent color in python icons
Summary:
BUG: 381051

FIXED-IN: 5.54

Add a separate, more recognizable icon for `application-x-python-bytecode`. Use a consistent, orange color for the `32px` and `64px` designs of `application-x-python-bytecode` and `text-x-python`.

The font used at all sizes except `64px` is made of custom shapes to be as pixel-perfect as possible.

Test Plan:
The new `application-x-python-bytecode` icons:
{F6514919}

Consistent style for `32px` and `64px` `application-x-python-bytecode` and `text-x-python`:
{F6489825}

Reviewers: #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17692