Add YaST and new preference icons
Summary:
Add YaST icons and a number of preference icons that will be useful for KDE software in the future.

Change `apps/32` to scalable so that the patch can pass the scalable test and add `apps/64` to `index.theme`

Test Plan:
`ViewBackground` is the background color that will be used in YaST Control Center.

Used by YaST. Most of these are symlinks to the new preference icons.
{F6513547}
{F6513548}

Not used by YaST, but currently included in the patch.
{F6513549}
{F6513550}

Reviewers: #vdg, cfeck, broulik, elvisangelaccio, ngraham

Reviewed By: #vdg, ngraham

Subscribers: rooty, sitter, fvogt, hellcp, ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17623