Improve contrast of preferences-system-time
Summary:
BUG: 390800

FIXED-IN: 5.55

Add outline to the `preferences-system-time` icon to improve contrast with the Breeze-dark desktop theme.

The outline is a bit small in the context menu, but that's due to this 32px icon being used at 16px size. Perhaps an action icon like `clock` should be used there instead.

Test Plan:
`preferences-system-time`:
{F6560737}

Contrast with tooltip:
{F6560738}

Contrast in context menu:
{F6560739}

Reviewers: #vdg, ngraham, ndavis

Reviewed By: #vdg, ngraham, ndavis

Subscribers: ndavis, ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D18432