Enforce new muted icon style in all action icons
Summary:
This patch updates the 16px `player-volume-muted` to be a symlink to the corresponding status icon. It also changes the 24px `audio-volume-muted` action icons to the new muted symbol introduced in {D17220}.

It also adjust the opacity of the inactive sound waves in the 24px `audio-volume-low` and `audio-volume-medium` icons to 0.25 to match the Breeze Plasma theme.

Reviewers: #vdg, ngraham, rooty

Reviewed By: #vdg, ngraham, rooty

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D19307