make start-here-kde also available start-here-kde-plasma
Summary:
the latter offers a less ambiguous name when we want to refer to something
plasma specific. a recent example is D20172 which defines the icon for
something related to the Plasma workspace, as such the start-here-kde icon
may not be the most appropriate across all icon themes (i.e. if the icon
is in fact the KDE logo it's not what we want). conversely an icon theme
cannot offer a real start-here-kde since we use it in place of a plasma
variant.

with the additional name we'd be free to use less ambiguous icon references
in the future.
also, because of how icon naming works themes that provide start-here-kde
but not start-here-kde-plasma would still fall back to start-here-kde

Reviewers: ngraham

Reviewed By: ngraham

Subscribers: broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20176