Use clockwise arrows for *refresh* and update-* icons
Summary:
Right now we are inconsistent with the arrow directionality for our circular icons:
some use clockwise arrows, and some use counterclockwise.

This patch standardizes everything on the clockwise direction. This not only makes
everything nice and consistent, but it also fixes the busy indictor spinner looking
weird (it rotates clockwise, like all opther busy spinners we've ever seen in any UI
toolkit).

BUG: 409914
FIXED-IN: 5.61

Test Plan:
{F7046657}

{F7046658}

{F7046660}

{F7046661}

{F7046668}

{F7046669}

{F7046670}

The eagle-eyed reviewer will note that the different icons also have different styles of arrows and their rotational offset is also different, but if that's a pre-existing issue that can be done in another patch, if it's also something we want to unify.

Reviewers: #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22534