Change Highlight to ButtonFocus in stylesheets
Summary: This fixes an issue where changing the selection background color in the colorscheme also changes the color of SVG elements using Highlight. This also prevents the elements from blending in with the selection background in colorschemes that use a different color for selection than the focus decoration.

Test Plan:
Before:
The blue SVG element blends in with the selection background.
{F7076720, size=full}

After:
In this screenshot, a darker blue is used for the selection background and a lighter blue is used for focus decorations. The icons use the focus decoration color.
{F7076713, size=full}

Reviewers: #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22647