Use the text rather than highlight color for rating/star emblem
Summary:
The problem with using the highlight color is that the emblem becomes invisible when
applied to a UI element that's selectable and (correctly) uses the default selection
color. Making it use the text color instead fixes the bug in all cases.

Test Plan:
Stars in various UIs (Discover, Dolphin) are now black rather than blue, and the icon
can now be used on buttons without becoming invisible.

Reviewers: ndavis, #vdg

Reviewed By: ndavis, #vdg

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24612