Add edit-opacity icons
Summary:
This patch adds edit-opacity icons requested by Okular for D15580.

I'm not quite sure about the naming, and even if the Qt SVG renderer can handle this transparency. I also considered a style with a checkerboard pattern, which is the symbol for transparency e.g. in GIMP, but I felt this is not recognizable enough to most people. If you search for a opacity/transparency icon, you often find something with two overlapping circles, however I think that would be too close to our path-intersection etc. icons.

I think this should be the last icon needed for https://bugs.kde.org/show_bug.cgi?id=408283, so I'm tagging this as closing that bug.

BUG: 408283
FIXED-IN: 5.64

Test Plan:
The new `edit-opacity` icons:
{F7624384}

Reviewers: #vdg, ngraham, ndavis

Reviewed By: #vdg, ngraham, ndavis

Subscribers: ndavis, ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24753