Fix incorrect rendering of 22px edit-opacity icon
Summary:
This patch should fix the incorrect rendering of the new 22px `edit-opacity` icon added in D24753.
The preview in Dolphin and Ikona no longer shows renders the file incorrectly, so should also resolve the issue for viewing in Okular and Cuttlefish.
While optimizing the file, `svgo` altered some paths, which apparently resulted in the file rendering incorrectly in the Qt SVG Renderer, but not in Inkscape.

Reviewers: #vdg, simgunz

Reviewed By: simgunz

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24807