Use a more correct and appropriate visual metaphor for "get hot new stuff" icon
Summary:
Right now it's a duplicate (not even a symlink) of the `favorite` icon, which is
semantically inaccurate. A more appropriate visual metaphor is "download", since that's
what GHNS will do. So this patch symlinks the GHNS icon to `edit-download`. This makes
it work for icons-only buttons as well, where the icon alone is sufficient to communicate
what will happen.

Thanks for @ndavis for suggesting this.

BUG: 400500
FIXED-IN: 5.64

Test Plan:
{F7651335, size=full}
{F7651336, size=full}

Reviewers: #vdg, leinir, ndavis

Reviewed By: #vdg, ndavis

Subscribers: kde-frameworks-devel, ndavis

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24871