Use an eyedropper for color-picker icons
Summary:
The current icon is a black droplet which doesn't really communicate "pick a color from
the screen", though that's what this icon is used for. Other apps use an eyedropper for
this too.

This icon is a modification of one submitted in https://bugs.kde.org/show_bug.cgi?id=403924.
I've cleaned it up to pixel-align everything, use the correct margins, and do the CSS stylesheet
magic.

BUG: 403924
FIXED-IN: 5.65

Test Plan:
Montage: {F7761511, size=full}
In Kolourpaint: {F7761515, size=full}

Reviewers: #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: ndavis, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25132