Change XHTML icon to be a purple HTML icon
Currently, XHTML is just using the XML icon, but XHTML is more of a "website" than an "XML file".
This patch changes it to use a purple HTML icon.
The purple still makes it somewhat "different" (it's XML) but focuses on the essence that it's
pretty much a webseite, too.

Differential Revision: https://phabricator.kde.org/D26397