Add status icons: data-error, data-warning, data-information
Summary:
These are generic icons to use to mark the status of data or other content.
The current available icons are bound to other context, like dialog or
cloud.

Use case of such icons would be e.g. with KDevelop to put such lines on the
marker border to mark lines with errors, warnings or hints as found by
parsers & checkers by a symbol, as well as in the problem report tool view
or as symbols in inline notes.
Those places use small icons, and best signal the type of marker by both
shape & color.

Reviewers: #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: #kdevelop, ngraham, davidre, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27272