breeze-icons autotests: skip symlinks in the builddir
Summary:
It's ok for symlinks in the builddir to be broken.
We can't (easily) detect that they actually point to something that will
exist after make install.

With a typical kdesrc-build setup the test passed because it ignored the
builddir (which is outside the srcdir) completely.

In CI however the builddir is a subdir, so the recursive directory
traversal in the unittest ends up there...

Test Plan: bin/symlink works again. I created a broken symlink in the srcdir, it detected it.

Reviewers: sitter, bshah, lbeltrame

Reviewed By: sitter

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28188