Fix some stylesheet colors in 2 icon files
Summary:
This fixes two tiny inconsistencies in existing icon files in icons/.
In the stylesheet they have #232627 for ColorScheme-Text.
Now they use #232629, like in all other icon files.

This inconsistency wasn’t a problem until now,
but would make it difficult to auto-generate icons-dark/ from icons.
The key in auto-generating is to substitute the colors in the stylesheets,
so the stylesheets need to have the correct colors.

Reviewers: #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28323