Use flag style iconography for view-calendar-holiday
Right now this icon re-uses the bookmarks icon, which is not really
semantically correct. Instead, let's use a flag-style icon, since
holidays often involve celebrations and national pride and things like
that, for which flags seem like more appropriate iconography.

Also make the 32px and larger versions use a colorful icon. Green is
chosen as the color because green is currently used as the color for
colorful calendar-based mime type icons.