Swap the system-upgrade icon for something more identifiable
The previous symlink pointed system-upgrade to go-up, which is an
incredibly generic icon which, in places where one might come across the
icon, seems quite entirely out of place. This patch swaps that for a
link to the update-none icon (so as not to colour the type of upgrade
available). It also removes the 16x16 variant, as there is no equivalent
update-none to link to.