Support old cmake
The current code for installation of icons breaks on older cmake,
in particular Debian/stable but also others, with:
    CMake Error at icons/CMakeLists.txt:29 (install):
      install FILES given directory "actions/16@2x" to install.
The code is refactored to go back to the installation method of
5.88 (installing the directories as is, which will also install the
symlinks), but still check for missing svg files in the icons-dark
and install them, too.

In contrast to 5.88, the list of directories to be installed is
dynamically generated and not hard-coded.