Remove monochrome versions of help-related app icons
App icons should always be colorful at all sizes. the help-browser and
system-help icons are in the Applications group, and currently violate
this by displaying monochrome versions at their 16px and 22px sizes.
For consistency, they should be removed.

The help-contents action icon currently uses symlinks to those icons for
its 16px and 22px monochrome versions (which are appropriate since it's
an action icon), so to avoid breaking it, those sizes are transformed
into files rather than being symlinks.

BUG: 456737
FIXED-IN: 5.97