Revert "Add SimpleScreenRecorder icon"
This reverts commit 498af3e6d57bab73bd0f3cca339c946a61d587e0.

This breaks the tray icon changing its appearance based on the
recording/not recording status of the app.

The Breeze version of this icon is not thaaaaaat different from the
original, so on balance it's probably not worth it if we can't fix the
issue, which is not easily possible.

BUG: 462905
CCBUG: 412490
FIXED-IN: 5.102