Delete false Gparted and Kwikdisk icons
They are simply clones of KDE Partition Manager's icon which is super
weird and confusing. If we want to theme 3rd-party apps, we need to be
respectful of their branding, not obliterate it in favor of our own.

Even with this commit, Partition Manager still suffers from this issue,
but that's because it doesn't have its own icon yet. We'll fix that,
then fix this issue for Partition Manager too.

BUG: 467319
FIXED-IN: 5.105