Add icons for showing and hiding a virtual keyboard
Right now we only have status icons for the virtual keyboard, and no
action icons. This means we can accurately show the current status, but
if we want an action icon that can be used for a button to show or hide
the virtual keyboard, we can't do that. With this commit, we can.


(cherry picked from commit f78855bff965f1234a5302e44e6809c8695bb9e2)