Add icons for showing and hiding a virtual keyboard
Right now we only have status icons for the virtual keyboard, and no
action icons. This means we can accurately show the current status, but
if we want an action icon that can be used for a button to show or hide
the virtual keyboard, we can't do that. With this commit, we can.