Increase moon size for redshift-status-on icon
After seeing this icon in a real system tray, it looks a bit too small next
to the other icons. The current version has a 1px margin at the left side
and the lower edge (to match up with a bit of the redshift-status-day icon).
The margins are in addition to the 3px ones that are required.

This commit increases the moon size to eliminate the extra 1px margins on
the left and at the bottom. The bulb & shine bits remains the same.