Remove test_scalable
It only applied to app icons for some reason and didn't seem to be
preventing any issues since none of the other icon folders had issues.

The patch description says that if searching for a 192x192 icon, no icon
would be found if there was no version of the icon in a scalable folder,
but that does not seem to be true. KIconTheme and QIconLoaderEngine will
still try to find a version of the icon equal to or less than the
specified size, even if you request a size that does not exist.

See
<https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html>
for the algorithm used to look up icons.