Revert "Add .gitattributes for svg file"
This reverts commit 48e5fc7152f5f38a7e31edbf5b05b9dc39f14e91.

Setting this attribute makes SVGs un-reviewable as text, which is a
deal-breaker for Breeze Icons where this is an important part of the
review process because we work hard to keep the text human-readable.

We will need to re-add this attribute again in the future only before
merging a merge request that is otherwise un-merge-able without it, and
then revert again.