Make 16 and 22px versions of input-touchscreen not always symbolic
We might actually want to use a scaled-down version of the colorful icon
for these small icon sizes--for example, in System Settings. To enable
that, rename the the 16px and 22px symbolic versions to have -symbolic
at the end, ensuring that we don't use them unless we actually asked for
an icon with that suffix.