Add zoom-in-map and zoom-out-map icons along with -symbolic versions
They are links to list-add and list-remove because we want them to look
like a simple + and - signs, which appears to be a widespread
convention in most GIS software. We don't want maps to have zoom icons
with a magnifying glass because a magnifying glass is already used for
search buttons/fields which are also fairly common in such apps.

Needed in Night Light KCM for
https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/4212