Add non-symbolic base version of network-wireless-bluetooth icons
Currently these icons only exist with the -symbolic suffix, meaning if
anything tries to access them without the suffix in the request, they
won't find it and will get back network-wireless instead, which looks
different and would constitute a visual bug.

Rename these icons to be named network-wireless-bluetooth and then add
symlinks for the -symbolic versions.