Make list-remove look like a red X
Currently its icon is a minus sign. This looks okay when paired with a
plus sign icon nearby, but suffers visually in other contexts and when
on a flat button. In these cases it struggles to look like an icon, and
can seem like a piece of punctuation or a visual glitch.

Instead, let's make this icon look like a red X. That's what the HIG
says we should use for icons that remove abstract content, too. See
https://develop.kde.org/hig/icons/#icons-for-destructive-actions.