Add Spinbox-specific decrease and increase icons
In the past, the Plasma spinbox used list-remove and list-add, relying
on these icons looking like minus and plus icons. But that was fragile
and wrong, and when one of the icons changed, the spinbox icons started
looking nonsensical.

Instead, let's add dedicated icons for this in the icon theme, and ask
the spinbox to use those. We can use the fallback functionality to use
the old icons names for themes without these.

CCBUG: 491312