Remove 48px draw-freehand symlink
It was symlinked to the Scribus app icon, which was not the right thing
to do. This was originally done to get a colorful icon in there at a
large size for an old version of Okular's sidebar which wanted to use
colorful icons, but that UI has been gone for a few years anyway.

BUG: 491887
FIXED-IN: 6.6