Add incomplete REUSE.toml
Instead of using .reuse/dep5, which is deprecated,
we should use the new official format.

Since I don't know who should be the "Upstream-Contact" (PackageSupplier),
I leave it as none.

I don't know how to handle the authorship of image files yet,
so I leave it as none.

This might not be necessary in the end. It's possible to add
SPDX headings to SVG files, like in
https://invent.kde.org/frameworks/breeze-icons/-/blob/master/icons/actions/16/folder-edit-sign-encrypt.svg?ref_type=heads&plain=1
(though it would be extremely painful to do so for every single file).