Delete 32px colorful folder action icon symlink
This causes the default folder icon to look different at large sizes
compared to e.g. folder-download.

What's happening is that the 32px-version of the default folder icon
is always used and scaled to larger sizes. This happens because there
is a link in the `actions` folder to the 32px folder icon (and only
for 32px!): `actions/32/folder.svg` links to `../../places/32/folder.svg`

This is unnecessary, and deleting the symlink fixes the problem.

BUG: 478493
FIXED-IN: 6.7