Monochromize not-explicitly-colorized symbolic folder icons
We have symbolic versions of folder icons, but currently they're all
just symlinks to the 16px and 22px versions of the base icon. Most of
the time this is fine.

Some of those folder icons are not monochrome at those sizes, though;
they intentionally retain their color. This is normally fine, but not
desired for symbolic icons, where we want to expunge as much color as
possible — except for the folder icons that are explicitly and
intentionally colored that way.

BUG: 494721
FIXED-IN: 6.8