Add boost and boost-boosted icons
These are used for boosts, reblogs, and retweets (whatever you want to call
them.) These are taken from Tokodon and originally made by Tusooa Zhu. I
modified them to recolor properly and made sure they fit within our icon
guidelines.

Since we already have a retweet icon, it's repurposed to look closer to this
icon. It still exists, but as a 16px of boost and of course to keep existing
applications working.