Remove all colorful 3rd-party app icons
When we override a 3rd party app's icon, we override the app's own
branding. This is generally a somewhat rude thing to do, and in many
cases we're violating their branding guidelines or trademarks. It also
puts us on the hook for updating these icons when the upstream branding
changes, which — let's be honest — we don't actually do.

Let's remove them.

For now the symbolic icons remain, since the explicit intention there
is to use a completely different style that blends into Plasma's System
Tray, so it's okay if it differs from the app's primary icon. However we
could remove these too if we want to be completely consistent here.

Fixes #96