Handle cross compiling for tools
This searches qrcAlias and generate-symbolic-dark in
the KF6_HOST_TOOLING path to avoid the complicated
double build structure required to compile them
for the build system when the host differs.

Also add option to install the host tools to have them for finding
later.

Done-With: Andre Heinecke <aheinecke@gnupg.org>